Add files via upload
I am starting a machine learning project series, this is the first project: Rock Vs Mine prediction model.
This machine learning model predicts based on user input if the input features are of a rock or a mine, I have used Python to code this and the libraries used are Sklearn, Numpy, and Pandas. The model used was Logistic Regression. The model accuracy in test data is 78%, which I look forward to improving in further projects.